requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
chardet==3.0.4
discord.py==1.0.1
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
typing-extensions==3.7.2
websockets==6.0
yarl==1.3.0
